sms14: merged TOP/cgra/ files
